Backend/requirements/base.txt
django-cors-headers==2.1.0
Django==1.11.15
djangorestframework==3.7.3
drf-nested-routers==0.91
django-filter==1.1.0
rest-condition==1.0.3
channels==1.1.8
PyMySQL==0.7.11
filelock==2.0.13
asgi-redis==1.4.3
cryptography==2.3
django-redis==4.8.0
lxml==4.1.1
requests==2.18.4
beautifulsoup4==4.6.0
html2text==2017.10.4
sqlparse==0.2.4
python-crontab==2.2.7
psutil==5.4.1
JPype1==0.6.3